Separate into partials, add team section
